refactor: remove dto import alias
