--- src/dto/request/__init__.py
@@ -1,6 +1,0 @@
-from .feedback import UserFeedbackRequest
-from .music import RecommendMusicRequest
-from .auth import ReLoginRequest
-
-
-__all__ = [UserFeedbackRequest, RecommendMusicRequest, ReLoginRequest]
