--- src/dto/response/__init__.py
@@ -1,4 +1,0 @@
-from .music import RecommendMusicResponse, RecommendMusic
-from .auth import SigninResponse, ReLoginResponse
-
-__all__ = [RecommendMusic, RecommendMusicResponse, SigninResponse, ReLoginResponse]
